chore: rename to tracegen and address comments
